Create ProducerDemoCallback.java
Callback is used to send message asynchronously along with information such as whether the data was correctly produced, where it was produced, about its offset and partition value.

For performing the callbacks, the user needs to implement a callback function. This function is implemented for asynchronously handling the request completion. That's why it's return type will be void. This function will be implemented in the block where the producer sends data to the Kafka. There is no requirement to make changes in other blocks of codes.

The callback function used by the producer is the onCompletion(). Basically, this method requires two arguments:

Metadata of the Record: Metadata of the record means fetching the information regarding the partition and its offsets. If it is not null, an error will be thrown.

Exception: There are following exceptions which can be thrown while processing:

1) Retriable exception: This exception says that the message may be sent.

2) Non-retriable exception: This exception throws the error that the message will never be sent.